fingerprint: switch to 32-bit hashes
32-bit hashes reduce the processing time of large ABIs, despite having
slightly more collisions.

PiperOrigin-RevId: 498390791

+++ fingerprint.h
@@ -33,7 +33,7 @@ namespace stg {
 //
 // Given any mutual dependencies between hashes, it falls back to a very poor
 // but safe hash for the affected nodes: the size of the SCC.
-std::unordered_map<Id, uint64_t> Fingerprint(
+std::unordered_map<Id, uint32_t> Fingerprint(
     const Graph& graph, Id root, Metrics& metrics);
 
 }  // namespace stg
